requirements.txt
asgiref==3.8.1
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.4
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==24.4.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crispy-bootstrap4==2024.10
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
django==4.2.21
django-allauth==65.9.0
django-crispy-forms==2.4
django-leaflet==0.28.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
python-dotenv==1.0.1
entrypoints==0.3
exceptiongroup==1.3.0
GDAL==3.0.4
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==2.1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
packaging==20.3
pexpect==4.6.0
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
pytest==8.3.5
pytest-django==4.11.1
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.8.2
soupsieve==2.7
sqlparse==0.5.3
ssh-import-id==5.10
systemd-python==234
tomli==2.2.1
tqdm==4.67.1
Twisted==18.9.0
typing-extensions==4.13.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wikipedia==1.4.0
zipp==1.0.0
zope.interface==4.7.1

# Novas dependências (v1.1 e v2.0)
firebase-admin==6.5.0
scikit-learn==1.5.2
scipy==1.14.1
